fig01_cifar_vs_mnist/requirements.txt
dill==0.3.6
GitPython==3.1.31
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.5.0
pycm==3.8
requests==2.28.2
ruamel.base==1.0.0
scikit_learn==1.2.2
scipy==1.10.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.65.0
